diabetes

+++ eval_ml.py
@@ -14,7 +14,7 @@ import matplotlib.pyplot as plt
 def separate_cat_and_num_cols(domain: Domain):
     # train_cols = [c for c in domain.attrs if c != target]
     train_cols_num = domain.get_numerical_cols()
-    train_cols_cat = domain.get_categorical_cols()+ domain.get_ordinal_cols()
+    train_cols_cat = domain.get_categorical_cols() + domain.get_ordinal_cols()
     return train_cols_num, train_cols_cat
 
 def get_Xy(domain: Domain, target, df_train: pd.DataFrame, df_test: pd.DataFrame,
@@ -29,8 +29,8 @@ def get_Xy(domain: Domain, target, df_train: pd.DataFrame, df_test: pd.DataFrame
     X_test = None
 
     if len(cols_cat) > 0:
-        X_cat_train = df_train[cols_cat].values
-        X_cat_test = df_test[cols_cat].values
+        X_cat_train = df_train[cols_cat].astype(int).values
+        X_cat_test = df_test[cols_cat].astype(int).values
         X = np.concatenate((X_cat_train, X_cat_test))
         # categories = [np.arange(domain[c]['size']) for c in cols_cat]
         categories = [np.unique(np.concatenate((df_train[c].values, df_test[c].values))) for c in cols_cat]
